+++ components.d.ts
@@ -5,9 +5,15 @@ import '@vue/runtime-core'
 
 declare module '@vue/runtime-core' {
   export interface GlobalComponents {
+    Footer: typeof import('./src/components/Layout/Footer.vue')['default']
     HelloWorld: typeof import('./src/components/HelloWorld.vue')['default']
+    HomeHero: typeof import('./src/components/Home/HomeHero.vue')['default']
+    HomeStoreTest: typeof import('./src/components/Home/HomeStoreTest.vue')['default']
+    Navbar: typeof import('./src/components/Layout/Navbar/index.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
+    StoreTest: typeof import('./src/components/Home/StoreTest.vue')['default']
+    ThemeChange: typeof import('./src/components/Layout/Navbar/ThemeChange.vue')['default']
   }
 }