+++ src/router/index.ts
@@ -1,0 +1,36 @@
+import { createRouter, createWebHistory, RouteRecordRaw } from 'vue-router';
+import NProgress from 'nprogress';
+
+import Index from '@/views/Index.vue';
+
+const routes: RouteRecordRaw[] = [
+  {
+    path: '/',
+    component: Index,
+    meta: {
+      title: 'Vite + Vue + TypeScript + Tailwind Starter Template',
+    },
+  },
+  {
+    path: '/home',
+    name: 'Home',
+    component: () => import('@/views/Home.vue'),
+  },
+];
+
+const router = createRouter({
+  history: createWebHistory(),
+  routes,
+});
+
+router.beforeEach((to, from) => {
+  if (!NProgress.isStarted()) {
+    NProgress.start();
+  }
+});
+
+router.afterEach((to, from) => {
+  NProgress.done();
+});
+
+export default router;