+++ src/views/Home.vue
@@ -1,0 +1,11 @@
+<template>
+  <Hero />
+  <StoreTest />
+  <div>Home 2</div>
+</template>
+
+<script setup lang="ts">
+import Hero from '@/components/Home/HomeHero.vue';
+import StoreTest from '@/components/Home/HomeStoreTest.vue';
+</script>
+<style scoped></style>